#18 Errors2String method moved to helper
