+++ src/helpers/CommonHelper.php
@@ -1,0 +1,23 @@
+<?php
+declare(strict_types = 1);
+
+namespace cusodede\permissions\helpers;
+
+/**
+ * Class CommonHelper
+ */
+class CommonHelper {
+	/**
+	 * @param array $errors
+	 * @param array|string $separator
+	 * @return string
+	 */
+	public static function Errors2String(array $errors, array|string $separator = "\n"):string {
+		$output = [];
+		foreach ($errors as $attribute => $attributeErrors) {
+			$error = is_array($attributeErrors)?implode($separator, $attributeErrors):$attributeErrors;
+			$output[] = "{$attribute}: {$error}";
+		}
+		return implode($separator, $output);
+	}
+}
